+++ src/io.c
@@ -145,3 +145,77 @@ int qb_find_occurrences(char *query, char *body){
     }
     return num_queries;
 }
+
+/*----------------------------------------------------------------------------//
+    QBINLINE
+//----------------------------------------------------------------------------*/
+
+void qb_replace_char_if_proceeding(char *verse, int verse_size,
+                                   char *prologue, int prologue_size,
+                                   char a, char b){
+
+    bool find_match = true;
+    bool match_found = false;
+    int count = 0;
+    for (int i = 0; i < verse_size; ++i){
+        while (find_match){
+            if (verse[i+count] != prologue[count]){
+                find_match = false;
+            }
+
+            count++;
+
+            if (count == prologue_size && find_match){
+                find_match = false;
+                match_found = true;
+            }
+        }
+
+        // match found
+        if (match_found &&
+            verse[i+count] == a){
+            verse[i+count] = b;
+            match_found = false;
+        }
+        find_match = true;
+        count = 0;
+    }
+}
+
+void qb_replace_char(char *verse, int verse_size, char a, char b){
+    for (int i = 0; i < verse_size; ++i){
+        if (verse[i] == a){
+            verse[i] = b;
+        }
+    }
+}
+
+void qb_preprocess_content(char *content){
+    if (qb_is_inlined(content)){
+        int content_size = strlen(content);
+        // replace all spaces by new lines
+        qb_replace_char(content, content_size, ' ', '\n');
+
+        // except for the arguments after some preprocessor options
+        // that need to be in the same line
+        qb_replace_char_if_proceeding(content, content_size,
+                                      "#ifdef", 7, '\n', ' ');
+        qb_replace_char_if_proceeding(content, content_size,
+                                      "#ifndef", 8, '\n', ' ');
+
+        // #define with two arguments will not work
+        qb_replace_char_if_proceeding(content, content_size,
+                                      "#define", 7, '\n', ' ');
+
+        // don't leave any spaces in arguments
+        qb_replace_char_if_proceeding(content, content_size,
+                                      "#if", 3, '\n', ' ');
+
+        // don't leave any spaces in arguments
+        qb_replace_char_if_proceeding(content, content_size,
+                                      "#elif", 5, '\n', ' ');
+        qb_replace_char_if_proceeding(content, content_size, "#pragma",
+                                      7, '\n', ' ');
+    }
+}
+
